refactor(script): split script.js into ES modules under tools/
Each tool (json, regex, base64, color, notes, url) now lives in its
own module file under tools/. Shared utilities (copyText, escapeHtml)
are in tools/utils.js. script.js becomes a thin entry point that
imports all modules. index.html updated to use type="module" which
also defers script execution until DOM is parsed.

Closes #4
