+++ index.html
@@ -436,6 +436,6 @@
     <p>Alpha &amp; Beta &mdash; Peer Developers &mdash; 2026</p>
   </footer>
 
-  <script src="script.js"></script>
+  <script type="module" src="script.js"></script>
 </body>
 </html>
