+++ tools/theme.js
@@ -1,0 +1,28 @@
+// Theme Toggle
+
+const themeToggle = document.getElementById('themeToggle');
+const html = document.documentElement;
+const THEME_KEY = 'peer-theme';
+
+function applyTheme(theme) {
+  html.setAttribute('data-theme', theme);
+  const isDark = theme === 'dark';
+  themeToggle.setAttribute('aria-label', isDark ? 'Switch to light mode' : 'Switch to dark mode');
+  themeToggle.querySelector('.theme-toggle__icon').textContent = isDark ? '☀' : '☾';
+}
+
+(function initTheme() {
+  const saved = localStorage.getItem(THEME_KEY);
+  const preferred = saved ?? (window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light');
+  applyTheme(preferred);
+})();
+
+themeToggle.addEventListener('click', () => {
+  const next = html.getAttribute('data-theme') === 'dark' ? 'light' : 'dark';
+  applyTheme(next);
+  localStorage.setItem(THEME_KEY, next);
+});
+
+window.matchMedia('(prefers-color-scheme: dark)').addEventListener('change', (e) => {
+  if (!localStorage.getItem(THEME_KEY)) applyTheme(e.matches ? 'dark' : 'light');
+});
